requirements.txt
discord.py>=2.3.0
python-dotenv
tortoise-orm
jishaku
git+https://github.com/Rapptz/discord-ext-menus
